fixed all -Wall bugs
Fixed all warnings the compiler started to emit when I added the -Wall flag to
the compilation command. The build should now be clean, making it easier to
spot actual errors.

+++ maxwell/ast/nodes/BinaryOpExpr.cpp
@@ -317,11 +317,11 @@ void BinaryOpExpr::decode(Decoder& d) {
 void BinaryOpExpr::updateHierarchyOfChildren() {
 	if (this->lhs) this->lhs->updateHierarchy(id + "lhs", repository, this);
 	if (this->rhs) this->rhs->updateHierarchy(id + "rhs", repository, this);
-	for (int i = 0; i < this->callArgs.size(); i++) {
+	for (unsigned i = 0; i < this->callArgs.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->callArgs[i]->updateHierarchy((id + "callArgs") + buf, repository, this);
 	}
-	for (int i = 0; i < this->callCandidates.size(); i++) {
+	for (unsigned i = 0; i < this->callCandidates.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->callCandidates[i]->updateHierarchy((id + "callCandidates") + buf, repository, this);
 	}
@@ -355,7 +355,7 @@ const NodePtr& BinaryOpExpr::resolvePath(const std::string& path) {
 					string idx_str = path.substr(9, dot);
 					int idx = atoi(idx_str.c_str());
 					const NodeVector& a = getCallArgs();
-					if (idx < 0 || idx >= a.size()) {
+					if (idx < 0 || idx >= (int)a.size()) {
 						throw std::runtime_error("Index into array '" + path.substr(0, 8) + "' is out of bounds.");
 					}
 					if (dot == string::npos) {
@@ -375,7 +375,7 @@ const NodePtr& BinaryOpExpr::resolvePath(const std::string& path) {
 					string idx_str = path.substr(15, dot);
 					int idx = atoi(idx_str.c_str());
 					const NodeVector& a = getCallCandidates();
-					if (idx < 0 || idx >= a.size()) {
+					if (idx < 0 || idx >= (int)a.size()) {
 						throw std::runtime_error("Index into array '" + path.substr(0, 14) + "' is out of bounds.");
 					}
 					if (dot == string::npos) {
