+++ maxwell/ast/nodes/CallExpr.cpp
@@ -315,15 +315,15 @@ void CallExpr::decode(Decoder& d) {
 
 void CallExpr::updateHierarchyOfChildren() {
 	if (this->context) this->context->updateHierarchy(id + "context", repository, this);
-	for (int i = 0; i < this->args.size(); i++) {
+	for (unsigned i = 0; i < this->args.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->args[i]->updateHierarchy((id + "args") + buf, repository, this);
 	}
-	for (int i = 0; i < this->callArgs.size(); i++) {
+	for (unsigned i = 0; i < this->callArgs.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->callArgs[i]->updateHierarchy((id + "callArgs") + buf, repository, this);
 	}
-	for (int i = 0; i < this->callCandidates.size(); i++) {
+	for (unsigned i = 0; i < this->callCandidates.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->callCandidates[i]->updateHierarchy((id + "callCandidates") + buf, repository, this);
 	}
@@ -357,7 +357,7 @@ const NodePtr& CallExpr::resolvePath(const std::string& path) {
 					string idx_str = path.substr(5, dot);
 					int idx = atoi(idx_str.c_str());
 					const NodeVector& a = getArgs();
-					if (idx < 0 || idx >= a.size()) {
+					if (idx < 0 || idx >= (int)a.size()) {
 						throw std::runtime_error("Index into array '" + path.substr(0, 4) + "' is out of bounds.");
 					}
 					if (dot == string::npos) {
@@ -382,7 +382,7 @@ const NodePtr& CallExpr::resolvePath(const std::string& path) {
 						string idx_str = path.substr(9, dot);
 						int idx = atoi(idx_str.c_str());
 						const NodeVector& a = getCallArgs();
-						if (idx < 0 || idx >= a.size()) {
+						if (idx < 0 || idx >= (int)a.size()) {
 							throw std::runtime_error("Index into array '" + path.substr(0, 8) + "' is out of bounds.");
 						}
 						if (dot == string::npos) {
@@ -402,7 +402,7 @@ const NodePtr& CallExpr::resolvePath(const std::string& path) {
 						string idx_str = path.substr(15, dot);
 						int idx = atoi(idx_str.c_str());
 						const NodeVector& a = getCallCandidates();
-						if (idx < 0 || idx >= a.size()) {
+						if (idx < 0 || idx >= (int)a.size()) {
 							throw std::runtime_error("Index into array '" + path.substr(0, 14) + "' is out of bounds.");
 						}
 						if (dot == string::npos) {
