+++ maxwell/ast/nodes/IfCaseExpr.cpp
@@ -200,7 +200,7 @@ void IfCaseExpr::decode(Decoder& d) {
 
 
 void IfCaseExpr::updateHierarchyOfChildren() {
-	for (int i = 0; i < this->conds.size(); i++) {
+	for (unsigned i = 0; i < this->conds.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->conds[i]->updateHierarchy((id + "conds") + buf, repository, this);
 	}
@@ -233,7 +233,7 @@ const NodePtr& IfCaseExpr::resolvePath(const std::string& path) {
 				string idx_str = path.substr(6, dot);
 				int idx = atoi(idx_str.c_str());
 				const NodeVector& a = getConds();
-				if (idx < 0 || idx >= a.size()) {
+				if (idx < 0 || idx >= (int)a.size()) {
 					throw std::runtime_error("Index into array '" + path.substr(0, 5) + "' is out of bounds.");
 				}
 				if (dot == string::npos) {
