+++ maxwell/ast/nodes/MemberAccessExpr.cpp
@@ -302,11 +302,11 @@ void MemberAccessExpr::updateHierarchyOfChildren() {
 	if (this->possibleType) this->possibleType->updateHierarchy(id + "possibleType", repository, this);
 	if (this->requiredType) this->requiredType->updateHierarchy(id + "requiredType", repository, this);
 	if (this->actualType) this->actualType->updateHierarchy(id + "actualType", repository, this);
-	for (int i = 0; i < this->callArgs.size(); i++) {
+	for (unsigned i = 0; i < this->callArgs.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->callArgs[i]->updateHierarchy((id + "callArgs") + buf, repository, this);
 	}
-	for (int i = 0; i < this->callCandidates.size(); i++) {
+	for (unsigned i = 0; i < this->callCandidates.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->callCandidates[i]->updateHierarchy((id + "callCandidates") + buf, repository, this);
 	}
@@ -337,7 +337,7 @@ const NodePtr& MemberAccessExpr::resolvePath(const std::string& path) {
 					string idx_str = path.substr(9, dot);
 					int idx = atoi(idx_str.c_str());
 					const NodeVector& a = getCallArgs();
-					if (idx < 0 || idx >= a.size()) {
+					if (idx < 0 || idx >= (int)a.size()) {
 						throw std::runtime_error("Index into array '" + path.substr(0, 8) + "' is out of bounds.");
 					}
 					if (dot == string::npos) {
@@ -357,7 +357,7 @@ const NodePtr& MemberAccessExpr::resolvePath(const std::string& path) {
 					string idx_str = path.substr(15, dot);
 					int idx = atoi(idx_str.c_str());
 					const NodeVector& a = getCallCandidates();
-					if (idx < 0 || idx >= a.size()) {
+					if (idx < 0 || idx >= (int)a.size()) {
 						throw std::runtime_error("Index into array '" + path.substr(0, 14) + "' is out of bounds.");
 					}
 					if (dot == string::npos) {
