+++ maxwell/ast/nodes/QualifiedTypeExpr.cpp
@@ -129,7 +129,7 @@ void QualifiedTypeExpr::decode(Decoder& d) {
 
 void QualifiedTypeExpr::updateHierarchyOfChildren() {
 	if (this->evaluatedType) this->evaluatedType->updateHierarchy(id + "evaluatedType", repository, this);
-	for (int i = 0; i < this->exprs.size(); i++) {
+	for (unsigned i = 0; i < this->exprs.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->exprs[i]->updateHierarchy((id + "exprs") + buf, repository, this);
 	}
@@ -160,7 +160,7 @@ const NodePtr& QualifiedTypeExpr::resolvePath(const std::string& path) {
 					string idx_str = path.substr(6, dot);
 					int idx = atoi(idx_str.c_str());
 					const NodeVector& a = getExprs();
-					if (idx < 0 || idx >= a.size()) {
+					if (idx < 0 || idx >= (int)a.size()) {
 						throw std::runtime_error("Index into array '" + path.substr(0, 5) + "' is out of bounds.");
 					}
 					if (dot == string::npos) {
