+++ maxwell/ast/nodes/TupleExpr.cpp
@@ -179,7 +179,7 @@ void TupleExpr::updateHierarchyOfChildren() {
 	if (this->possibleType) this->possibleType->updateHierarchy(id + "possibleType", repository, this);
 	if (this->requiredType) this->requiredType->updateHierarchy(id + "requiredType", repository, this);
 	if (this->actualType) this->actualType->updateHierarchy(id + "actualType", repository, this);
-	for (int i = 0; i < this->args.size(); i++) {
+	for (unsigned i = 0; i < this->args.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->args[i]->updateHierarchy((id + "args") + buf, repository, this);
 	}
@@ -210,7 +210,7 @@ const NodePtr& TupleExpr::resolvePath(const std::string& path) {
 					string idx_str = path.substr(5, dot);
 					int idx = atoi(idx_str.c_str());
 					const NodeVector& a = getArgs();
-					if (idx < 0 || idx >= a.size()) {
+					if (idx < 0 || idx >= (int)a.size()) {
 						throw std::runtime_error("Index into array '" + path.substr(0, 4) + "' is out of bounds.");
 					}
 					if (dot == string::npos) {
