+++ maxwell/ast/nodes/UnaryOpExpr.cpp
@@ -313,11 +313,11 @@ void UnaryOpExpr::decode(Decoder& d) {
 
 void UnaryOpExpr::updateHierarchyOfChildren() {
 	if (this->expr) this->expr->updateHierarchy(id + "expr", repository, this);
-	for (int i = 0; i < this->callArgs.size(); i++) {
+	for (unsigned i = 0; i < this->callArgs.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->callArgs[i]->updateHierarchy((id + "callArgs") + buf, repository, this);
 	}
-	for (int i = 0; i < this->callCandidates.size(); i++) {
+	for (unsigned i = 0; i < this->callCandidates.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->callCandidates[i]->updateHierarchy((id + "callCandidates") + buf, repository, this);
 	}
@@ -351,7 +351,7 @@ const NodePtr& UnaryOpExpr::resolvePath(const std::string& path) {
 					string idx_str = path.substr(9, dot);
 					int idx = atoi(idx_str.c_str());
 					const NodeVector& a = getCallArgs();
-					if (idx < 0 || idx >= a.size()) {
+					if (idx < 0 || idx >= (int)a.size()) {
 						throw std::runtime_error("Index into array '" + path.substr(0, 8) + "' is out of bounds.");
 					}
 					if (dot == string::npos) {
@@ -371,7 +371,7 @@ const NodePtr& UnaryOpExpr::resolvePath(const std::string& path) {
 					string idx_str = path.substr(15, dot);
 					int idx = atoi(idx_str.c_str());
 					const NodeVector& a = getCallCandidates();
-					if (idx < 0 || idx >= a.size()) {
+					if (idx < 0 || idx >= (int)a.size()) {
 						throw std::runtime_error("Index into array '" + path.substr(0, 14) + "' is out of bounds.");
 					}
 					if (dot == string::npos) {
