+++ maxwell/ast/nodes/UnionType.cpp
@@ -63,7 +63,7 @@ void UnionType::decode(Decoder& d) {
 
 
 void UnionType::updateHierarchyOfChildren() {
-	for (int i = 0; i < this->types.size(); i++) {
+	for (unsigned i = 0; i < this->types.size(); i++) {
 		char buf[32]; snprintf(buf, 31, "%i", i);
 		this->types[i]->updateHierarchy((id + "types") + buf, repository, this);
 	}
@@ -81,7 +81,7 @@ const NodePtr& UnionType::resolvePath(const std::string& path) {
 			string idx_str = path.substr(6, dot);
 			int idx = atoi(idx_str.c_str());
 			const NodeVector& a = getTypes();
-			if (idx < 0 || idx >= a.size()) {
+			if (idx < 0 || idx >= (int)a.size()) {
 				throw std::runtime_error("Index into array '" + path.substr(0, 5) + "' is out of bounds.");
 			}
 			if (dot == string::npos) {
